pyproject.toml
[project]
name = "youtube-video-summary"
version = "0.1.0"
description = "MCP server for fetching YouTube video transcripts and generating customizable summaries"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.0.2",
    "youtube-transcript-api>=1.2.3",
    "pyyaml>=6.0",
]
